test(mitm-addon): centralize usage timer scaffolding
